requirements.txt
flask
flask-login
gunicorn
SQLAlchemy
psycopg2-binary
argon2-cffi
